+++ tests/phpunit/tests/admin/wpRollbackAutoUpdate.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Test WP_Rollback_Auto_Update class.
+ * Test WP_Rollback_Auto_Update.
  *
  * @package WP_Rollback_Auto_Update
  *
@@ -8,5 +8,8 @@
  * @group plugins
  */
 
+/**
+ * Tests_Admin_WpRollbackAutoUpdate
+ */
 class Tests_Admin_WpRollbackAutoUpdate extends WP_UnitTestCase {
 }