+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 [unreleased]
 
+#### 5.0.3 / 2023-03-22
+* update @since
+* update using constant to check version for when `move_dir()` was committed
+* update using constant to check version for when `Rollback` was committed
+* update for PR compatibility
+* developery stuff
+
 #### 5.0.2 / 2023-02-05
 * make variables static to retain value during auto-update run